restructured for MCP
restructured to work with mcp server
added mcp server
