+++ VS/duHastApplications/roommate/Cargo.toml
@@ -17,3 +17,4 @@ anyhow = "1"
 csv = "1"
 futures-util = "0.3"
 tokio-util = { version = "0.7", features = ["io"] }
+rmcp = { version = "2.1.0", features = ["server", "macros", "transport-io", "schemars"] }
